requirements.txt
flask==3.0.0
twilio==9.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
openpyxl==3.1.2

# Optional: Only needed if USE_LLM=true
# openai>=1.12.0
# httpx>=0.24.0
# anthropic==0.18.1
